moved things around to avoid circular includes
