+++ stan/math/prim/fun/to_ref.hpp
@@ -44,21 +44,6 @@ inline ref_type_t<T&&> to_ref_if(T&& a) {
   return std::forward<T>(a);
 }
 
-#ifdef STAN_OPENCL
-/**
- * Converts given Eigen expression into one that can be directly copied to an
- * OpenCL device to create `matrix_cl`. If given expression can be directly
- * copied, this is a no-op that should be optimized away.
- * @tparam T argument type
- * @param a argument
- * @return optionally evaluated argument
- */
-template <typename T>
-inline ref_type_for_opencl_t<T&&> to_ref_for_opencl(T&& a) {
-  return std::forward<T>(a);
-}
-#endif
-
 }  // namespace math
 }  // namespace stan
 #endif
